Add variables to let juceaide be compiled as a native executable

+++ modules/juce_events/messages/juce_MessageManager.cpp
@@ -28,7 +28,7 @@ MessageManager::MessageManager() noexcept
 {
     if (JUCEApplicationBase::isStandaloneApp())
         Thread::setCurrentThreadName ("JUCE Message Thread");
-    DBG("messageThreadId " << (int)messageThreadId);
+    DBG("messageThreadId " << (intptr_t)messageThreadId);
 }
 
 MessageManager::~MessageManager() noexcept
